backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-environ>=0.11,<1.0
django-cors-headers>=4.4,<5.0
django-ratelimit>=4.1,<5.0
Pillow>=10.0,<12.0
bleach>=6.1,<7.0
meilisearch>=0.33,<1.0
